Hooked up modals to room connection
