+++ src/pages/room/RoomContainer.tsx
@@ -8,11 +8,7 @@ import { useModals } from "@components/Modals/useModals";
 import { DynamicScene } from "./DynamicScene";
 import { roomModals } from "./Modals";
 import { RoomBottom } from "./RoomBottom";
-import {
-  useRoomContextValue,
-  RoomContext,
-  RoomContextSettings,
-} from "./RoomContext";
+import { RoomContextSettings } from "./RoomContext";
 import { RoomHead } from "./RoomHead";
 
 export type RoomContainerProps = {
@@ -20,7 +16,6 @@ export type RoomContainerProps = {
 };
 
 export const RoomContainer: React.FC<RoomContainerProps> = ({ settings }) => {
-  const roomContextValue = useRoomContextValue(settings);
   const [modalNodes, setModal] = useModals(roomModals);
 
   useEffect(() => {
@@ -28,8 +23,7 @@ export const RoomContainer: React.FC<RoomContainerProps> = ({ settings }) => {
   }, []);
 
   return (
-    <RoomContext.Provider value={roomContextValue}>
-      <DynamicScene />
+    <DynamicScene settings={settings}>
       <div id={rootElementId}>
         <BaseHead title={`${settings.room} |`} />
       </div>
@@ -37,6 +31,6 @@ export const RoomContainer: React.FC<RoomContainerProps> = ({ settings }) => {
       <RoomHead setModal={setModal} />
       <RoomBottom setModal={setModal} />
       {modalNodes}
-    </RoomContext.Provider>
+    </DynamicScene>
   );
 };
